feat(client): remove icons from stats to reduce cluttering
